requirements.txt
torchvision
torch
librosa
transformers
faiss-gpu
numpy<2.0.0
opencv-python